feat: anilist subcommands independent files
